requirements.txt
tensorflow-gpu==2.4.1
scikit-learn
imgaug==0.4.0
opencv-python==4.5.1.48
pixellib==0.5.6
trimesh==3.9.8
pyrender==0.1.45
pyrealsense2
labelme
pysimplegui
numpy==1.19.2
tqdm
psutil
pandas
pyopengl-accelerate
klampt==0.8.7